refactor model loading
